api/requirements.txt
# ============================================================================
# SANTANDER CREDIT RISK PLATFORM - REQUIREMENTS
# Versão 2.2.1 - Corrigida e Validada
# ============================================================================

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.10.0
pydantic-settings==2.6.0

# Machine Learning
lightgbm==4.5.0
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
joblib==1.4.2

# Security & Authentication
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4  # Password hashing
python-multipart==0.0.12  # OAuth2 form data
cryptography==43.0.3  # Data encryption

# Logging & Monitoring
structlog==24.4.0
prometheus-client==0.21.0
python-json-logger==3.1.0

# Database
psycopg2-binary==2.9.10  # PostgreSQL
sqlalchemy==2.0.36
alembic==1.14.0

# Cache
redis==5.2.0
hiredis==3.0.0

# HTTP & Networking
httpx==0.27.2
requests==2.32.3

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
email-validator==2.2.0  # Email validation

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
black==24.10.0
flake8==7.1.1
mypy==1.13.0

# Rate Limiting
slowapi==0.1.9

# Validação de CPF
validate-docbr==1.10.0

# Parquet support
pyarrow==18.0.0
bcrypt==4.2.0

# ============================================================================
# NOTAS DE INSTALAÇÃO
# ============================================================================
#
# Instalação completa:
#   pip install -r requirements.txt
#
# Compatibilidade:
#   Python 3.8 - 3.12
#   Testado em: Ubuntu 22.04, Windows 11, macOS 14
#
# ============================================================================
tabulate==0.9.0
